+++ providers/class-two-factor-fido-u2f-admin.php
@@ -2,6 +2,12 @@
 /**
  * Class for registering & modifying FIDO U2F security keys.
  *
+ * @package Two_Factor
+ */
+
+/**
+ * Class for registering & modifying FIDO U2F security keys.
+ *
  * @since 0.1-dev
  *
  * @package Two_Factor
@@ -24,13 +30,13 @@ class Two_Factor_FIDO_U2F_Admin {
 	 * @static
 	 */
 	public static function add_hooks() {
-		add_action( 'admin_enqueue_scripts',       array( __CLASS__, 'enqueue_assets' ) );
+		add_action( 'admin_enqueue_scripts', array( __CLASS__, 'enqueue_assets' ) );
 		add_action( 'show_user_security_settings', array( __CLASS__, 'show_user_profile' ) );
-		add_action( 'personal_options_update',     array( __CLASS__, 'catch_submission' ), 0 );
-		add_action( 'edit_user_profile_update',    array( __CLASS__, 'catch_submission' ), 0 );
-		add_action( 'load-profile.php',            array( __CLASS__, 'catch_delete_security_key' ) );
-		add_action( 'load-user-edit.php',          array( __CLASS__, 'catch_delete_security_key' ) );
-		add_action( 'wp_ajax_inline-save-key',     array( __CLASS__, 'wp_ajax_inline_save' ) );
+		add_action( 'personal_options_update', array( __CLASS__, 'catch_submission' ), 0 );
+		add_action( 'edit_user_profile_update', array( __CLASS__, 'catch_submission' ), 0 );
+		add_action( 'load-profile.php', array( __CLASS__, 'catch_delete_security_key' ) );
+		add_action( 'load-user-edit.php', array( __CLASS__, 'catch_delete_security_key' ) );
+		add_action( 'wp_ajax_inline-save-key', array( __CLASS__, 'wp_ajax_inline_save' ) );
 	}
 
 	/**
@@ -44,16 +50,20 @@ class Two_Factor_FIDO_U2F_Admin {
 	 * @param string $hook Current page.
 	 */
 	public static function enqueue_assets( $hook ) {
-		if ( ! in_array( $hook, array( 'user-edit.php', 'profile.php' ) ) ) {
+		if ( ! in_array( $hook, array( 'user-edit.php', 'profile.php' ), true ) ) {
 			return;
 		}
 
-		$user_id = get_current_user_id();
+		$user_id = Two_Factor_Core::current_user_being_edited();
+		if ( ! $user_id ) {
+			return;
+		}
+
 		$security_keys = Two_Factor_FIDO_U2F::get_security_keys( $user_id );
 
-		// @todo Ensure that scripts don't fail because of missing u2fL10n
+		// @todo Ensure that scripts don't fail because of missing u2fL10n.
 		try {
-			$data = Two_Factor_FIDO_U2F::$u2f->getRegisterData( $security_keys );
+			$data              = Two_Factor_FIDO_U2F::$u2f->getRegisterData( $security_keys );
 			list( $req,$sigs ) = $data;
 
 			update_user_meta( $user_id, self::REGISTER_DATA_USER_META_KEY, $req );
@@ -65,14 +75,14 @@ class Two_Factor_FIDO_U2F_Admin {
 			'fido-u2f-admin',
 			plugins_url( 'css/fido-u2f-admin.css', __FILE__ ),
 			null,
-			'0.1.0-dev.1'
+			self::asset_version()
 		);
 
 		wp_enqueue_script(
 			'fido-u2f-admin',
 			plugins_url( 'js/fido-u2f-admin.js', __FILE__ ),
 			array( 'jquery', 'fido-u2f-api' ),
-			'0.1.0-dev.3',
+			self::asset_version(),
 			true
 		);
 
@@ -81,14 +91,15 @@ class Two_Factor_FIDO_U2F_Admin {
 		 */
 
 		$translation_array = array(
+			'user_id'  => $user_id,
 			'register' => array(
 				'request' => $req,
-				'sigs' => $sigs,
+				'sigs'    => $sigs,
 			),
-			'text' => array(
-				'insert' => esc_html__( 'Now insert (and tap) your Security Key.', 'two-factor' ),
-				'error' => esc_html__( 'U2F request failed.', 'two-factor' ),
-				'error_codes' => array(
+			'text'     => array(
+				'insert'            => esc_html__( 'Now insert (and tap) your Security Key.', 'two-factor' ),
+				'error'             => esc_html__( 'U2F request failed.', 'two-factor' ),
+				'error_codes'       => array(
 					// Map u2f.ErrorCodes to error messages.
 					0 => esc_html__( 'Request OK.', 'two-factor' ),
 					1 => esc_html__( 'Other U2F error.', 'two-factor' ),
@@ -97,7 +108,7 @@ class Two_Factor_FIDO_U2F_Admin {
 					4 => esc_html__( 'U2F device ineligible.', 'two-factor' ),
 					5 => esc_html__( 'U2F request timeout reached.', 'two-factor' ),
 				),
-				'u2f_not_supported' => esc_html__( 'FIDO U2F is not supported in your web browser. Try using Google Chrome.', 'two-factor' ),
+				'u2f_not_supported' => esc_html__( 'FIDO U2F appears to be not supported by your web browser. Try using Google Chrome or Firefox.', 'two-factor' ),
 			),
 		);
 
@@ -115,7 +126,7 @@ class Two_Factor_FIDO_U2F_Admin {
 			'inline-edit-key',
 			plugins_url( 'js/fido-u2f-admin-inline-edit.js', __FILE__ ),
 			array( 'jquery' ),
-			'0.1.0-dev.1',
+			self::asset_version(),
 			true
 		);
 
@@ -126,6 +137,18 @@ class Two_Factor_FIDO_U2F_Admin {
 				'error' => esc_html__( 'Error while saving the changes.', 'two-factor' ),
 			)
 		);
+	}
+
+	/**
+	 * Return the current asset version number.
+	 *
+	 * Added as own helper to allow swapping the implementation once we inject
+	 * it as a dependency.
+	 *
+	 * @return string
+	 */
+	protected static function asset_version() {
+		return Two_Factor_FIDO_U2F::asset_version();
 	}
 
 	/**
@@ -185,8 +208,8 @@ class Two_Factor_FIDO_U2F_Admin {
 			<p><a href="https://support.google.com/accounts/answer/6103523"><?php esc_html_e( 'You can find FIDO U2F Security Key devices for sale from here.', 'two-factor' ); ?></a></p>
 
 			<?php
-				require( TWO_FACTOR_DIR . 'providers/class.two-factor-fido-u2f-admin-list-table.php' );
-				$u2f_list_table = new Two_Factor_FIDO_U2F_Admin_List_Table();
+				require TWO_FACTOR_DIR . 'providers/class-two-factor-fido-u2f-admin-list-table.php';
+				$u2f_list_table        = new Two_Factor_FIDO_U2F_Admin_List_Table();
 				$u2f_list_table->items = $security_keys;
 				$u2f_list_table->prepare_items();
 				$u2f_list_table->display();
@@ -215,7 +238,7 @@ class Two_Factor_FIDO_U2F_Admin {
 
 			try {
 				$response = json_decode( stripslashes( $_POST['u2f_response'] ) );
-				$reg = Two_Factor_FIDO_U2F::$u2f->doRegister( get_user_meta( $user_id, self::REGISTER_DATA_USER_META_KEY, true ), $response );
+				$reg      = Two_Factor_FIDO_U2F::$u2f->doRegister( get_user_meta( $user_id, self::REGISTER_DATA_USER_META_KEY, true ), $response );
 				$reg->new = true;
 
 				Two_Factor_FIDO_U2F::add_security_key( $user_id, $reg );
@@ -225,9 +248,14 @@ class Two_Factor_FIDO_U2F_Admin {
 
 			delete_user_meta( $user_id, self::REGISTER_DATA_USER_META_KEY );
 
-			wp_safe_redirect( add_query_arg( array(
-				'new_app_pass' => 1,
-			), wp_get_referer() ) . '#security-keys-section' );
+			wp_safe_redirect(
+				add_query_arg(
+					array(
+						'new_app_pass' => 1,
+					),
+					wp_get_referer()
+				) . '#security-keys-section'
+			);
 			exit;
 		}
 	}
@@ -243,9 +271,11 @@ class Two_Factor_FIDO_U2F_Admin {
 	 * @static
 	 */
 	public static function catch_delete_security_key() {
-		$user_id = get_current_user_id();
-		if ( ! empty( $_REQUEST['delete_security_key'] ) ) {
+		$user_id = Two_Factor_Core::current_user_being_edited();
+
+		if ( ! empty( $user_id ) && ! empty( $_REQUEST['delete_security_key'] ) ) {
 			$slug = $_REQUEST['delete_security_key'];
+
 			check_admin_referer( "delete_security_key-{$slug}", '_nonce_delete_security_key' );
 
 			Two_Factor_FIDO_U2F::delete_security_key( $user_id, $slug );
@@ -281,7 +311,7 @@ class Two_Factor_FIDO_U2F_Admin {
 	 * @return string
 	 */
 	public static function delete_link( $item ) {
-		$delete_link = add_query_arg( 'delete_security_key', $item->keyHandle );
+		$delete_link = add_query_arg( 'delete_security_key', $item->keyHandle ); // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
 		$delete_link = wp_nonce_url( $delete_link, "delete_security_key-{$item->keyHandle}", '_nonce_delete_security_key' );
 		return sprintf( '<a href="%1$s">%2$s</a>', esc_url( $delete_link ), esc_html__( 'Delete', 'two-factor' ) );
 	}
@@ -297,22 +327,21 @@ class Two_Factor_FIDO_U2F_Admin {
 	public static function wp_ajax_inline_save() {
 		check_ajax_referer( 'keyinlineeditnonce', '_inline_edit' );
 
-		require( TWO_FACTOR_DIR . 'providers/class.two-factor-fido-u2f-admin-list-table.php' );
+		require TWO_FACTOR_DIR . 'providers/class-two-factor-fido-u2f-admin-list-table.php';
 		$wp_list_table = new Two_Factor_FIDO_U2F_Admin_List_Table();
 
 		if ( ! isset( $_POST['keyHandle'] ) ) {
 			wp_die();
 		}
 
-		$user_id = get_current_user_id();
-
+		$user_id       = Two_Factor_Core::current_user_being_edited();
 		$security_keys = Two_Factor_FIDO_U2F::get_security_keys( $user_id );
 		if ( ! $security_keys ) {
 			wp_die();
 		}
 
 		foreach ( $security_keys as &$key ) {
-			if ( $key->keyHandle === $_POST['keyHandle'] ) {
+			if ( $key->keyHandle === $_POST['keyHandle'] ) { // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
 				break;
 			}
 		}
